Fix fatal ninja-delete edge case
This occured when the command comment was deleted in between calling `validate_comment(comment)`
and `reply(comment)`. The _make_comment call on line 57 was unprotected by a `try`, and when
comment.author.name was attempted (line 137) it threw an error, as `comment.author == None`
I moved the `_make_message` call inside the already existing `try`, which will now
catch this edge case.